refactor: Change imports from auth_api to auth_service
