+++ bigfastapi/settings.py
@@ -7,7 +7,7 @@ from fastapi.param_functions import Depends
 from sqlalchemy.exc import IntegrityError
 
 from bigfastapi.db.database import get_db
-from .auth_api import is_authenticated
+from bigfastapi.services.auth_service import is_authenticated
 from .models import organization_models as _models
 from .models import settings_models as models
 from .schemas import settings_schemas as schemas
